RestaurantTest.java
package com.RestaurantFinder;

import org.junit.jupiter.api.Test;

import java.time.LocalTime;
import java.util.ArrayList;
import static org.junit.jupiter.api.Assertions.*;


class RestaurantTest {
        Restaurant restaurant;

        //spoof variable acts as the menu selected by the user
        ArrayList<Item> spoof = new ArrayList<Item>();

        public void restaurantCreation(){
            LocalTime openingTime = LocalTime.parse("10:30:00");
            LocalTime closingTime = LocalTime.parse("22:00:00");
            restaurant = new Restaurant("Amelie's cafe","Chennai",openingTime,closingTime);
            restaurant.addToMenu("Sweet corn soup",119);
            restaurant.addToMenu("Vegetable lasagne", 269);
        }
        //>>>>>>>>>>>>>>>>>>>>>>>>>OPEN/CLOSED<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
        @Test
        public void is_restaurant_open_should_return_true_if_time_is_between_opening_and_closing_time(){
            restaurantCreation();
            restaurant.setClosingTime(LocalTime.now().plusMinutes(10));
            assertTrue(restaurant.isRestaurantOpen());
        }

        @Test
        public void is_restaurant_open_should_return_false_if_time_is_outside_opening_and_closing_time(){
            restaurantCreation();
            restaurant.setClosingTime(LocalTime.now().minusMinutes(10));
            assertFalse(restaurant.isRestaurantOpen());
        }
        //<<<<<<<<<<<<<<<<<<<<<<<<<OPEN/CLOSED>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>

        //>>>>>>>>>>>>>>>>>>>>>>>>>>>ORDER VALUE<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
        @Test
        public void order_value_should_get_cumulative_total_when_collection_of_items_selected(){
            restaurantCreation();
            spoof = restaurant.getMenu();
            assertEquals(506,restaurant.getOrderValue(spoof));
        }

        @Test
        public void order_value_should_reduce_cumulative_total_when_an_item_removed(){
            restaurantCreation();
            spoof = restaurant.getMenu();
            int total = restaurant.getOrderValue(spoof);
            int afterTotal = spoof.get(1).getPrice();
            spoof.remove(1);
            assertEquals(total-afterTotal,restaurant.getOrderValue(spoof));
        }
        //<<<<<<<<<<<<<<<<<<<<<<<<<ORDER VALUE>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>

        //>>>>>>>>>>>>>>>>>>>>>>>>>>>MENU<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
        @Test
        public void adding_item_to_menu_should_increase_menu_size_by_1(){
            restaurantCreation();
            int initialMenuSize = restaurant.getMenu().size();
            restaurant.addToMenu("Sizzling brownie",319);
            assertEquals(initialMenuSize+1,restaurant.getMenu().size());
        }
        @Test
        public void removing_item_from_menu_should_decrease_menu_size_by_1() throws itemNotFoundException {
            restaurantCreation();
            int initialMenuSize = restaurant.getMenu().size();
            restaurant.removeFromMenu("Vegetable lasagne");
            assertEquals(initialMenuSize-1,restaurant.getMenu().size());
        }
        @Test
        public void removing_item_that_does_not_exist_should_throw_exception() {
            restaurantCreation();
            assertThrows(itemNotFoundException.class,
                    ()->restaurant.removeFromMenu("French fries"));
        }

        private class List<T> {
        }
        //<<<<<<<<<<<<<<<<<<<<<<<MENU>>>>>>>>>>>>>>>>>>>>>>>>>>>>>

    }

